Rename injector loader
