+++ src/main/java/org/hydev/mcpm/client/arguments/InteractorFactory.java
@@ -8,12 +8,12 @@ import org.hydev.mcpm.client.database.tracker.PluginTracker;
 import org.hydev.mcpm.client.display.progress.ProgressBarFetcherListener;
 import org.hydev.mcpm.client.export.ExportInteractor;
 import org.hydev.mcpm.client.export.ExportPluginsBoundary;
-import org.hydev.mcpm.client.injector.LoadBoundary;
-import org.hydev.mcpm.client.injector.LocalJarBoundary;
-import org.hydev.mcpm.client.injector.LocalJarFinder;
-import org.hydev.mcpm.client.injector.PluginLoader;
-import org.hydev.mcpm.client.injector.ReloadBoundary;
-import org.hydev.mcpm.client.injector.UnloadBoundary;
+import org.hydev.mcpm.client.loader.LoadBoundary;
+import org.hydev.mcpm.client.loader.LocalJarBoundary;
+import org.hydev.mcpm.client.loader.LocalJarFinder;
+import org.hydev.mcpm.client.loader.PluginLoader;
+import org.hydev.mcpm.client.loader.ReloadBoundary;
+import org.hydev.mcpm.client.loader.UnloadBoundary;
 import org.hydev.mcpm.client.installer.InstallBoundary;
 import org.hydev.mcpm.client.installer.InstallInteractor;
 import org.hydev.mcpm.client.installer.PluginDownloader;
