+++ src/main/java/org/hydev/mcpm/client/arguments/InteractorFactoryBoundary.java
@@ -5,10 +5,10 @@ import org.hydev.mcpm.client.database.fetcher.DatabaseFetcherListener;
 import org.hydev.mcpm.client.database.mirrors.MirrorSelectBoundary;
 import org.hydev.mcpm.client.database.tracker.PluginTracker;
 import org.hydev.mcpm.client.export.ExportPluginsBoundary;
-import org.hydev.mcpm.client.injector.LoadBoundary;
-import org.hydev.mcpm.client.injector.LocalJarBoundary;
-import org.hydev.mcpm.client.injector.ReloadBoundary;
-import org.hydev.mcpm.client.injector.UnloadBoundary;
+import org.hydev.mcpm.client.loader.LoadBoundary;
+import org.hydev.mcpm.client.loader.LocalJarBoundary;
+import org.hydev.mcpm.client.loader.ReloadBoundary;
+import org.hydev.mcpm.client.loader.UnloadBoundary;
 import org.hydev.mcpm.client.installer.InstallBoundary;
 import org.hydev.mcpm.client.installer.PluginDownloader;
 import org.hydev.mcpm.client.list.ListAllBoundary;
