+++ src/main/java/org/hydev/mcpm/client/commands/controllers/UnloadController.java
@@ -1,7 +1,7 @@
 package org.hydev.mcpm.client.commands.controllers;
 
-import org.hydev.mcpm.client.injector.PluginNotFoundException;
-import org.hydev.mcpm.client.injector.UnloadBoundary;
+import org.hydev.mcpm.client.loader.PluginNotFoundException;
+import org.hydev.mcpm.client.loader.UnloadBoundary;
 
 import java.util.List;
 import java.util.function.Consumer;
