+++ src/main/java/org/hydev/mcpm/client/installer/InstallInteractor.java
@@ -3,21 +3,22 @@ package org.hydev.mcpm.client.installer;
 import org.hydev.mcpm.client.Downloader;
 import org.hydev.mcpm.client.database.tracker.PluginTracker;
 import org.hydev.mcpm.client.display.progress.ProgressBarFetcherListener;
-import org.hydev.mcpm.client.injector.PluginLoader;
+import org.hydev.mcpm.client.installer.output.InstallResult;
+import org.hydev.mcpm.client.loader.LoadBoundary;
+import org.hydev.mcpm.client.loader.PluginLoader;
+import org.hydev.mcpm.client.loader.PluginNotFoundException;
 import org.hydev.mcpm.client.local.LocalDatabaseFetcher;
 import org.hydev.mcpm.client.local.LocalPluginTracker;
+import org.hydev.mcpm.client.search.SearchPackagesBoundary;
+import org.hydev.mcpm.client.search.SearchPackagesInput;
 import org.hydev.mcpm.client.search.SearchPackagesType;
 import org.hydev.mcpm.client.search.SearchPackagesResult;
 import org.hydev.mcpm.client.search.SearchInteractor;
-import org.hydev.mcpm.client.installer.output.InstallResult;
-import org.hydev.mcpm.client.search.*;
-import org.hydev.mcpm.client.injector.LoadBoundary;
-import org.hydev.mcpm.client.injector.PluginNotFoundException;
-import org.hydev.mcpm.client.installer.output.InstallResult.Type;
 import org.hydev.mcpm.client.installer.input.InstallInput;
 import org.hydev.mcpm.client.display.presenters.InstallPresenter;
 import org.hydev.mcpm.client.models.PluginModel;
 import org.hydev.mcpm.utils.ColorLogger;
+import org.hydev.mcpm.client.installer.output.InstallResult.Type;
 
 import java.io.File;
 import java.net.URI;
