+++ src/main/java/org/hydev/mcpm/client/loader/LocalJarBoundary.java
@@ -1,4 +1,4 @@
-package org.hydev.mcpm.client.injector;
+package org.hydev.mcpm.client.loader;
 
 import java.io.File;
